refactor(one-trust): Move cookie setting code
